engine: extract esm/pointing.js
