+++ engine/esm/transpiled.js
@@ -750,47 +750,6 @@ set_makeNewHipsProperties(function (imageset) {
 });
 
 
-// wwtlib.Pointing
-
-export function Pointing() {
-  this.theta = 0;
-  this.phi = 0;
-}
-
-Pointing.create = function (theta, phi) {
-  var temp = new Pointing();
-  temp.theta = theta;
-  temp.phi = phi;
-  return temp;
-};
-
-var Pointing$ = {
-  normalizeTheta: function () {
-    this.theta = HealpixUtils.fmodulo(this.theta, 2 * Math.PI);
-    if (this.theta > Math.PI) {
-      this.phi += Math.PI;
-      this.theta = 2 * Math.PI - this.theta;
-    }
-  },
-
-  normalize: function () {
-    this.normalizeTheta();
-    this.phi = HealpixUtils.fmodulo(this.phi, 2 * Math.PI);
-  },
-
-  toString: function () {
-    var s = new ss.StringBuilder();
-    s.append('ptg(');
-    s.append(this.theta);
-    s.append(',');
-    s.append(this.phi);
-    s.append(')');
-    return s.toString();
-  }
-};
-
-registerType("Pointing", [Pointing, Pointing$, null]);
-
 // wwtlib.Folder
 // requires: Place, Imageset, Tour, FolderUp
